Updated get projects part
